+++ src/main/kotlin/com/github/fcramer/gradle/jasperreports/tasks/CompilationParameters.kt
@@ -3,7 +3,9 @@ package com.github.fcramer.gradle.jasperreports.tasks
 import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.provider.Property
+import org.gradle.api.provider.SetProperty
 import org.gradle.workers.WorkParameters
+import java.io.File
 
 interface CompilationParameters : WorkParameters {
     val source: RegularFileProperty
@@ -13,4 +15,6 @@ interface CompilationParameters : WorkParameters {
     val validateXml: Property<Boolean>
     val keepJava: Property<Boolean>
     val tmpDir: DirectoryProperty
+
+    val classpath: SetProperty<File>
 }