+++ src/main/kotlin/com/github/fcramer/gradle/jasperreports/tasks/JasperReportsCompileWork.kt
@@ -14,6 +14,7 @@ abstract class JasperReportsCompileWork : WorkAction<CompilationParameters> {
         val validateXml = parameters.validateXml.get()
         val keepJava = parameters.keepJava.get()
         val tmpDir = parameters.tmpDir.asFile.get()
+        val classpath = parameters.classpath.get()
 
         val task = CompilationTask(
             source,
@@ -23,6 +24,7 @@ abstract class JasperReportsCompileWork : WorkAction<CompilationParameters> {
                 validateXml,
                 keepJava,
                 tmpDir,
+                classpath,
             ),
         )