+++ src/main/kotlin/com/github/fcramer/gradle/jasperreports/utils/CompilationUtils.kt
@@ -7,6 +7,7 @@ import net.sf.jasperreports.engine.JasperReportsContext
 import net.sf.jasperreports.engine.SimpleJasperReportsContext
 import net.sf.jasperreports.engine.design.JRCompiler
 import net.sf.jasperreports.engine.xml.JRReportSaxParserFactory
+import java.io.File
 
 fun compileReport(task: CompilationTask) {
     val context = configureJasperReportsContext(task.configuration)
@@ -16,6 +17,10 @@ fun compileReport(task: CompilationTask) {
 
 private fun configureJasperReportsContext(configuration: TaskConfiguration): JasperReportsContext {
     val context = SimpleJasperReportsContext()
+    context.setProperty(
+        JRCompiler.COMPILER_CLASSPATH,
+        configuration.classpath.joinToString(separator = File.pathSeparator) { it.absolutePath },
+    )
     context.setProperty(JRReportSaxParserFactory.COMPILER_XML_VALIDATION, configuration.isValidateXml.toString())
     configuration.compiler?.let { context.setProperty(JRCompiler.COMPILER_PREFIX, it) }
     context.setProperty(JRCompiler.COMPILER_KEEP_JAVA_FILE, configuration.isKeepJava.toString())